mpi/comm: split MPIR implementations from MPI functions
Move all MPIR implementation code to src/mpi/comm/comm_impl.c.

We are going to separate the implementation code from the mpi top level
source, so PMPI_LOCAL cannot be used any more.
