+++ src/mpi/comm/comm_create_group.c
@@ -25,94 +25,6 @@ int MPI_Comm_create_group(MPI_Comm comm, MPI_Group group, int tag, MPI_Comm * ne
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Comm_create_group
 #define MPI_Comm_create_group PMPI_Comm_create_group
-
-/* comm create group impl; assumes that the standard error checking
- * has already taken place in the calling function */
-int MPIR_Comm_create_group(MPIR_Comm * comm_ptr, MPIR_Group * group_ptr, int tag,
-                           MPIR_Comm ** newcomm_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Context_id_t new_context_id = 0;
-    int *mapping = NULL;
-    int n;
-
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIR_COMM_CREATE_GROUP);
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIR_COMM_CREATE_GROUP);
-
-    MPIR_Assert(comm_ptr->comm_kind == MPIR_COMM_KIND__INTRACOMM);
-
-    n = group_ptr->size;
-    *newcomm_ptr = NULL;
-
-    /* Shift tag into the tagged coll space */
-    tag |= MPIR_TAG_COLL_BIT;
-
-    /* Create a new communicator from the specified group members */
-
-    if (group_ptr->rank != MPI_UNDEFINED) {
-        MPIR_Comm *mapping_comm = NULL;
-
-        /* For this routine, creation of the id is collective over the input
-         *group*, so processes not in the group do not participate. */
-
-        mpi_errno = MPIR_Get_contextid_sparse_group(comm_ptr, group_ptr, tag, &new_context_id, 0);
-        MPIR_ERR_CHECK(mpi_errno);
-        MPIR_Assert(new_context_id != 0);
-
-        mpi_errno = MPII_Comm_create_calculate_mapping(group_ptr, comm_ptr,
-                                                       &mapping, &mapping_comm);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        /* Get the new communicator structure and context id */
-
-        mpi_errno = MPIR_Comm_create(newcomm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        (*newcomm_ptr)->recvcontext_id = new_context_id;
-        (*newcomm_ptr)->rank = group_ptr->rank;
-        (*newcomm_ptr)->comm_kind = comm_ptr->comm_kind;
-        /* Since the group has been provided, let the new communicator know
-         * about the group */
-        (*newcomm_ptr)->local_comm = 0;
-        (*newcomm_ptr)->local_group = group_ptr;
-        MPIR_Group_add_ref(group_ptr);
-
-        (*newcomm_ptr)->remote_group = group_ptr;
-        MPIR_Group_add_ref(group_ptr);
-        (*newcomm_ptr)->context_id = (*newcomm_ptr)->recvcontext_id;
-        (*newcomm_ptr)->remote_size = (*newcomm_ptr)->local_size = n;
-
-        /* Setup the communicator's vc table.  This is for the remote group,
-         * which is the same as the local group for intracommunicators */
-        mpi_errno = MPII_Comm_create_map(n, 0, mapping, NULL, mapping_comm, *newcomm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        (*newcomm_ptr)->tainted = comm_ptr->tainted;
-        mpi_errno = MPIR_Comm_commit(*newcomm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else {
-        /* This process is not in the group */
-        new_context_id = 0;
-    }
-
-  fn_exit:
-    MPL_free(mapping);
-
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIR_COMM_CREATE_GROUP);
-    return mpi_errno;
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-    if (*newcomm_ptr != NULL) {
-        MPIR_Comm_release(*newcomm_ptr);
-        new_context_id = 0;     /* MPIR_Comm_release frees the new ctx id */
-    }
-    if (new_context_id != 0)
-        MPIR_Free_contextid(new_context_id);
-    /* --END ERROR HANDLING-- */
-
-    goto fn_exit;
-}
-
 #endif /* !defined(MPICH_MPI_FROM_PMPI) */
 
 /*@
