+++ src/mpi/comm/comm_split_type.c
@@ -24,268 +24,6 @@ int MPI_Comm_split_type(MPI_Comm comm, int split_type, int key, MPI_Info info, M
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Comm_split_type
 #define MPI_Comm_split_type PMPI_Comm_split_type
-
-static const char *SHMEM_INFO_KEY = "shmem_topo";
-
-static int node_split(MPIR_Comm * comm_ptr, int key, const char *hint_str,
-                      MPIR_Comm ** newcomm_ptr);
-static int compare_info_hint(const char *hint_str, MPIR_Comm * comm_ptr, int *info_args_are_equal);
-
-int MPIR_Comm_split_type(MPIR_Comm * user_comm_ptr, int split_type, int key,
-                         MPIR_Info * info_ptr, MPIR_Comm ** newcomm_ptr)
-{
-    MPIR_Comm *comm_ptr = NULL;
-    int mpi_errno = MPI_SUCCESS;
-
-    /* split out the undefined processes */
-    mpi_errno = MPIR_Comm_split_impl(user_comm_ptr, split_type == MPI_UNDEFINED ? MPI_UNDEFINED : 0,
-                                     key, &comm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (split_type == MPI_UNDEFINED) {
-        *newcomm_ptr = NULL;
-        goto fn_exit;
-    }
-
-    if (split_type == MPI_COMM_TYPE_SHARED) {
-        mpi_errno = MPIR_Comm_split_type_self(comm_ptr, split_type, key, newcomm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else if (split_type == MPIX_COMM_TYPE_NEIGHBORHOOD) {
-        mpi_errno =
-            MPIR_Comm_split_type_neighborhood(comm_ptr, split_type, key, info_ptr, newcomm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else {
-        MPIR_ERR_SETANDJUMP(mpi_errno, MPI_ERR_ARG, "**arg");
-    }
-
-  fn_exit:
-    if (comm_ptr)
-        MPIR_Comm_free_impl(comm_ptr);
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Comm_split_type_impl(MPIR_Comm * comm_ptr, int split_type, int key,
-                              MPIR_Info * info_ptr, MPIR_Comm ** newcomm_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    /* Only MPI_COMM_TYPE_SHARED, MPI_UNDEFINED, and
-     * NEIGHBORHOOD are supported */
-    MPIR_Assert(split_type == MPI_COMM_TYPE_SHARED ||
-                split_type == MPI_UNDEFINED || split_type == MPIX_COMM_TYPE_NEIGHBORHOOD);
-
-    if (MPIR_Comm_fns != NULL && MPIR_Comm_fns->split_type != NULL) {
-        mpi_errno = MPIR_Comm_fns->split_type(comm_ptr, split_type, key, info_ptr, newcomm_ptr);
-    } else {
-        mpi_errno = MPIR_Comm_split_type(comm_ptr, split_type, key, info_ptr, newcomm_ptr);
-    }
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIR_Comm_set_info_impl(*newcomm_ptr, info_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Comm_split_type_self(MPIR_Comm * user_comm_ptr, int split_type, int key,
-                              MPIR_Comm ** newcomm_ptr)
-{
-    MPIR_Comm *comm_ptr = NULL;
-    MPIR_Comm *comm_self_ptr;
-    int mpi_errno = MPI_SUCCESS;
-
-    /* split out the undefined processes */
-    mpi_errno = MPIR_Comm_split_impl(user_comm_ptr, split_type == MPI_UNDEFINED ? MPI_UNDEFINED : 0,
-                                     key, &comm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (split_type == MPI_UNDEFINED) {
-        *newcomm_ptr = NULL;
-        goto fn_exit;
-    }
-
-    MPIR_Comm_get_ptr(MPI_COMM_SELF, comm_self_ptr);
-    mpi_errno = MPIR_Comm_dup_impl(comm_self_ptr, NULL, newcomm_ptr);
-
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    if (comm_ptr)
-        MPIR_Comm_free_impl(comm_ptr);
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Comm_split_type_by_node(MPIR_Comm * user_comm_ptr, int split_type, int key,
-                                 MPIR_Comm ** newcomm_ptr)
-{
-    MPIR_Comm *comm_ptr = NULL;
-    int mpi_errno = MPI_SUCCESS;
-    int color;
-
-    /* split out the undefined processes */
-    mpi_errno = MPIR_Comm_split_impl(user_comm_ptr, split_type == MPI_UNDEFINED ? MPI_UNDEFINED : 0,
-                                     key, &comm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (split_type == MPI_UNDEFINED) {
-        *newcomm_ptr = NULL;
-        goto fn_exit;
-    }
-
-    mpi_errno = MPID_Get_node_id(comm_ptr, comm_ptr->rank, &color);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIR_Comm_split_impl(comm_ptr, color, key, newcomm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    if (comm_ptr)
-        MPIR_Comm_free_impl(comm_ptr);
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Comm_split_type_node_topo(MPIR_Comm * user_comm_ptr, int split_type, int key,
-                                   MPIR_Info * info_ptr, MPIR_Comm ** newcomm_ptr)
-{
-    MPIR_Comm *comm_ptr;
-    int mpi_errno = MPI_SUCCESS;
-    int flag = 0;
-    char hint_str[MPI_MAX_INFO_VAL + 1];
-    int info_args_are_equal;
-    *newcomm_ptr = NULL;
-
-    mpi_errno = MPIR_Comm_split_type_by_node(user_comm_ptr, split_type, key, &comm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (comm_ptr == NULL) {
-        MPIR_Assert(split_type == MPI_UNDEFINED);
-        *newcomm_ptr = NULL;
-        goto fn_exit;
-    }
-
-    if (info_ptr) {
-        MPIR_Info_get_impl(info_ptr, SHMEM_INFO_KEY, MPI_MAX_INFO_VAL, hint_str, &flag);
-    }
-
-    if (!flag) {
-        hint_str[0] = '\0';
-    }
-
-    mpi_errno = compare_info_hint(hint_str, comm_ptr, &info_args_are_equal);
-
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /* if all processes do not have the same hint_str, skip
-     * topology-aware comm split */
-    if (!info_args_are_equal)
-        goto use_node_comm;
-
-    /* if no info key is given, skip topology-aware comm split */
-    if (!info_ptr)
-        goto use_node_comm;
-
-    /* if hw topology is not initialized, skip topology-aware comm split */
-    if (!MPIR_hwtopo_is_initialized())
-        goto use_node_comm;
-
-    if (flag) {
-        mpi_errno = node_split(comm_ptr, key, hint_str, newcomm_ptr);
-
-        MPIR_ERR_CHECK(mpi_errno);
-
-        MPIR_Comm_free_impl(comm_ptr);
-
-        goto fn_exit;
-    }
-
-  use_node_comm:
-    *newcomm_ptr = comm_ptr;
-
-  fn_exit:
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-static int node_split(MPIR_Comm * comm_ptr, int key, const char *hint_str, MPIR_Comm ** newcomm_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_hwtopo_gid_t gid;
-
-    gid = MPIR_hwtopo_get_obj_by_name(hint_str);
-
-    mpi_errno = MPIR_Comm_split_impl(comm_ptr, gid, key, newcomm_ptr);
-
-    return mpi_errno;
-}
-
-static int compare_info_hint(const char *hint_str, MPIR_Comm * comm_ptr, int *info_args_are_equal)
-{
-    int hint_str_size = strlen(hint_str);
-    int hint_str_size_max;
-    int hint_str_equal;
-    int hint_str_equal_global = 0;
-    char *hint_str_global = NULL;
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Errflag_t errflag = MPIR_ERR_NONE;
-
-    /* Find the maximum hint_str size.  Each process locally compares
-     * its hint_str size to the global max, and makes sure that this
-     * comparison is successful on all processes. */
-    mpi_errno =
-        MPID_Allreduce(&hint_str_size, &hint_str_size_max, 1, MPI_INT, MPI_MAX, comm_ptr, &errflag);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    hint_str_equal = (hint_str_size == hint_str_size_max);
-
-    mpi_errno =
-        MPID_Allreduce(&hint_str_equal, &hint_str_equal_global, 1, MPI_INT, MPI_LAND,
-                       comm_ptr, &errflag);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (!hint_str_equal_global)
-        goto fn_exit;
-
-
-    /* Now that the sizes of the hint_strs match, check to make sure
-     * the actual hint_strs themselves are the equal */
-    hint_str_global = (char *) MPL_malloc(strlen(hint_str), MPL_MEM_OTHER);
-
-    mpi_errno =
-        MPID_Allreduce(hint_str, hint_str_global, strlen(hint_str), MPI_CHAR,
-                       MPI_MAX, comm_ptr, &errflag);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    hint_str_equal = !memcmp(hint_str, hint_str_global, strlen(hint_str));
-
-    mpi_errno =
-        MPID_Allreduce(&hint_str_equal, &hint_str_equal_global, 1, MPI_INT, MPI_LAND,
-                       comm_ptr, &errflag);
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    MPL_free(hint_str_global);
-
-    *info_args_are_equal = hint_str_equal_global;
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif /* MPICH_MPI_FROM_PMPI */
 
 /*@
