+++ src/mpi/comm/intercomm_create.c
@@ -25,126 +25,6 @@ int MPI_Intercomm_create(MPI_Comm local_comm, int local_leader, MPI_Comm peer_co
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Intercomm_create
 #define MPI_Intercomm_create PMPI_Intercomm_create
-
-int MPIR_Intercomm_create_impl(MPIR_Comm * local_comm_ptr, int local_leader,
-                               MPIR_Comm * peer_comm_ptr, int remote_leader, int tag,
-                               MPIR_Comm ** new_intercomm_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Context_id_t final_context_id, recvcontext_id;
-    int remote_size = 0, *remote_lpids = NULL;
-    int comm_info[3];
-    int is_low_group = 0;
-    MPIR_Errflag_t errflag = MPIR_ERR_NONE;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIR_COMM_KIND__INTERCOMM_CREATE_IMPL);
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIR_COMM_KIND__INTERCOMM_CREATE_IMPL);
-
-    /* Shift tag into the tagged coll space */
-    tag |= MPIR_TAG_COLL_BIT;
-
-    mpi_errno = MPID_Intercomm_exchange_map(local_comm_ptr, local_leader,
-                                            peer_comm_ptr, remote_leader,
-                                            &remote_size, &remote_lpids, &is_low_group);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /*
-     * Create the contexts.  Each group will have a context for sending
-     * to the other group. All processes must be involved.  Because
-     * we know that the local and remote groups are disjoint, this
-     * step will complete
-     */
-    MPL_DBG_MSG_FMT(MPIR_DBG_COMM, VERBOSE,
-                    (MPL_DBG_FDEST, "About to get contextid (local_size=%d) on rank %d",
-                     local_comm_ptr->local_size, local_comm_ptr->rank));
-    /* In the multi-threaded case, MPIR_Get_contextid_sparse assumes that the
-     * calling routine already holds the single criticial section */
-    /* TODO: Make sure this is tag-safe */
-    mpi_errno = MPIR_Get_contextid_sparse(local_comm_ptr, &recvcontext_id, FALSE);
-    MPIR_ERR_CHECK(mpi_errno);
-    MPIR_Assert(recvcontext_id != 0);
-    MPL_DBG_MSG_FMT(MPIR_DBG_COMM, VERBOSE, (MPL_DBG_FDEST, "Got contextid=%d", recvcontext_id));
-
-    /* Leaders can now swap context ids and then broadcast the value
-     * to the local group of processes */
-    if (local_comm_ptr->rank == local_leader) {
-        MPIR_Context_id_t remote_context_id;
-
-        mpi_errno =
-            MPIC_Sendrecv(&recvcontext_id, 1, MPIR_CONTEXT_ID_T_DATATYPE, remote_leader, tag,
-                          &remote_context_id, 1, MPIR_CONTEXT_ID_T_DATATYPE, remote_leader, tag,
-                          peer_comm_ptr, MPI_STATUS_IGNORE, &errflag);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        final_context_id = remote_context_id;
-
-        /* Now, send all of our local processes the remote_lpids,
-         * along with the final context id */
-        comm_info[0] = final_context_id;
-        MPL_DBG_MSG(MPIR_DBG_COMM, VERBOSE, "About to bcast on local_comm");
-        mpi_errno = MPIR_Bcast(comm_info, 1, MPI_INT, local_leader, local_comm_ptr, &errflag);
-        MPIR_ERR_CHECK(mpi_errno);
-        MPIR_ERR_CHKANDJUMP(errflag, mpi_errno, MPI_ERR_OTHER, "**coll_fail");
-        MPL_DBG_MSG_D(MPIR_DBG_COMM, VERBOSE, "end of bcast on local_comm of size %d",
-                      local_comm_ptr->local_size);
-    } else {
-        /* we're the other processes */
-        MPL_DBG_MSG(MPIR_DBG_COMM, VERBOSE, "About to receive bcast on local_comm");
-        mpi_errno = MPIR_Bcast(comm_info, 1, MPI_INT, local_leader, local_comm_ptr, &errflag);
-        MPIR_ERR_CHECK(mpi_errno);
-        MPIR_ERR_CHKANDJUMP(errflag, mpi_errno, MPI_ERR_OTHER, "**coll_fail");
-
-        /* Extract the context and group sign informatin */
-        final_context_id = comm_info[0];
-    }
-
-    /* At last, we now have the information that we need to build the
-     * intercommunicator */
-
-    /* All processes in the local_comm now build the communicator */
-
-    mpi_errno = MPIR_Comm_create(new_intercomm_ptr);
-    if (mpi_errno)
-        goto fn_fail;
-
-    (*new_intercomm_ptr)->context_id = final_context_id;
-    (*new_intercomm_ptr)->recvcontext_id = recvcontext_id;
-    (*new_intercomm_ptr)->remote_size = remote_size;
-    (*new_intercomm_ptr)->local_size = local_comm_ptr->local_size;
-    (*new_intercomm_ptr)->rank = local_comm_ptr->rank;
-    (*new_intercomm_ptr)->comm_kind = MPIR_COMM_KIND__INTERCOMM;
-    (*new_intercomm_ptr)->local_comm = 0;
-    (*new_intercomm_ptr)->is_low_group = is_low_group;
-
-    mpi_errno = MPID_Create_intercomm_from_lpids(*new_intercomm_ptr, remote_size, remote_lpids);
-    if (mpi_errno)
-        goto fn_fail;
-
-    MPIR_Comm_map_dup(*new_intercomm_ptr, local_comm_ptr, MPIR_COMM_MAP_DIR__L2L);
-
-    /* Inherit the error handler (if any) */
-    MPID_THREAD_CS_ENTER(POBJ, MPIR_THREAD_POBJ_COMM_MUTEX(local_comm_ptr));
-    (*new_intercomm_ptr)->errhandler = local_comm_ptr->errhandler;
-    if (local_comm_ptr->errhandler) {
-        MPIR_Errhandler_add_ref(local_comm_ptr->errhandler);
-    }
-    MPID_THREAD_CS_EXIT(POBJ, MPIR_THREAD_POBJ_COMM_MUTEX(local_comm_ptr));
-
-    (*new_intercomm_ptr)->tainted = 1;
-    mpi_errno = MPIR_Comm_commit(*new_intercomm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-
-  fn_exit:
-    MPL_free(remote_lpids);
-    remote_lpids = NULL;
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIR_COMM_KIND__INTERCOMM_CREATE_IMPL);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-
 #endif /* MPICH_MPI_FROM_PMPI */
