+++ src/mpi/comm/intercomm_merge.c
@@ -25,166 +25,6 @@ int MPI_Intercomm_merge(MPI_Comm intercomm, int high, MPI_Comm * newintracomm)
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Intercomm_merge
 #define MPI_Intercomm_merge PMPI_Intercomm_merge
-
-/* This function creates mapping for new communicator
- * basing on network addresses of existing communicator.
- */
-
-static int create_and_map(MPIR_Comm * comm_ptr, int local_high, MPIR_Comm * new_intracomm_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int i;
-
-    /* Now we know which group comes first.  Build the new mapping
-     * from the existing comm */
-    if (local_high) {
-        /* remote group first */
-        MPIR_Comm_map_dup(new_intracomm_ptr, comm_ptr, MPIR_COMM_MAP_DIR__R2L);
-
-        MPIR_Comm_map_dup(new_intracomm_ptr, comm_ptr, MPIR_COMM_MAP_DIR__L2L);
-        for (i = 0; i < comm_ptr->local_size; i++)
-            if (i == comm_ptr->rank)
-                new_intracomm_ptr->rank = comm_ptr->remote_size + i;
-    } else {
-        /* local group first */
-        MPIR_Comm_map_dup(new_intracomm_ptr, comm_ptr, MPIR_COMM_MAP_DIR__L2L);
-        for (i = 0; i < comm_ptr->local_size; i++)
-            if (i == comm_ptr->rank)
-                new_intracomm_ptr->rank = i;
-
-        MPIR_Comm_map_dup(new_intracomm_ptr, comm_ptr, MPIR_COMM_MAP_DIR__R2L);
-    }
-
-    return mpi_errno;
-}
-
-int MPIR_Intercomm_merge_impl(MPIR_Comm * comm_ptr, int high, MPIR_Comm ** new_intracomm_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int local_high, remote_high, new_size;
-    MPIR_Context_id_t new_context_id;
-    MPIR_Errflag_t errflag = MPIR_ERR_NONE;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIR_COMM_KIND__INTERCOMM_MERGE_IMPL);
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIR_COMM_KIND__INTERCOMM_MERGE_IMPL);
-    /* Make sure that we have a local intercommunicator */
-    if (!comm_ptr->local_comm) {
-        /* Manufacture the local communicator */
-        mpi_errno = MPII_Setup_intercomm_localcomm(comm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-    }
-
-    /* Find the "high" value of the other group of processes.  This
-     * will be used to determine which group is ordered first in
-     * the generated communicator.  high is logical */
-    local_high = high;
-    if (comm_ptr->rank == 0) {
-        /* This routine allows use to use the collective communication
-         * context rather than the point-to-point context. */
-        mpi_errno = MPIC_Sendrecv(&local_high, 1, MPI_INT, 0, 0,
-                                  &remote_high, 1, MPI_INT, 0, 0, comm_ptr,
-                                  MPI_STATUS_IGNORE, &errflag);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        /* If local_high and remote_high are the same, then order is arbitrary.
-         * we use the is_low_group in the intercomm in this case. */
-        MPL_DBG_MSG_FMT(MPIR_DBG_COMM, VERBOSE,
-                        (MPL_DBG_FDEST, "local_high=%d remote_high=%d is_low_group=%d",
-                         local_high, remote_high, comm_ptr->is_low_group));
-        if (local_high == remote_high) {
-            local_high = !(comm_ptr->is_low_group);
-        }
-    }
-
-    /*
-     * All processes in the local group now need to get the
-     * value of local_high, which may have changed if both groups
-     * of processes had the same value for high
-     */
-    mpi_errno = MPIR_Bcast(&local_high, 1, MPI_INT, 0, comm_ptr->local_comm, &errflag);
-    MPIR_ERR_CHECK(mpi_errno);
-    MPIR_ERR_CHKANDJUMP(errflag, mpi_errno, MPI_ERR_OTHER, "**coll_fail");
-
-    /*
-     * For the intracomm, we need a consistent context id.
-     * That means that one of the two groups needs to use
-     * the recvcontext_id and the other must use the context_id
-     * The recvcontext_id is unique on each process, but another
-     * communicator may use the context_id. Therefore, we do a small hack.
-     * We set both flags indicating a sub-communicator (intra-node and
-     * inter-node) to one. This is normally not possible (a communicator
-     * is either intra- or inter-node) - which makes this context_id unique.
-     *
-     */
-
-    new_size = comm_ptr->local_size + comm_ptr->remote_size;
-
-    mpi_errno = MPIR_Comm_create(new_intracomm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (local_high) {
-        (*new_intracomm_ptr)->context_id =
-            MPIR_CONTEXT_SET_FIELD(SUBCOMM, comm_ptr->recvcontext_id, 3);
-    } else {
-        (*new_intracomm_ptr)->context_id = MPIR_CONTEXT_SET_FIELD(SUBCOMM, comm_ptr->context_id, 3);
-    }
-    (*new_intracomm_ptr)->recvcontext_id = (*new_intracomm_ptr)->context_id;
-    (*new_intracomm_ptr)->remote_size = (*new_intracomm_ptr)->local_size = new_size;
-    (*new_intracomm_ptr)->rank = -1;
-    (*new_intracomm_ptr)->comm_kind = MPIR_COMM_KIND__INTRACOMM;
-
-    /* Now we know which group comes first.  Build the new mapping
-     * from the existing comm */
-    mpi_errno = create_and_map(comm_ptr, local_high, (*new_intracomm_ptr));
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /* We've setup a temporary context id, based on the context id
-     * used by the intercomm.  This allows us to perform the allreduce
-     * operations within the context id algorithm, since we already
-     * have a valid (almost - see comm_create_hook) communicator.
-     */
-    (*new_intracomm_ptr)->tainted = 1;
-    mpi_errno = MPIR_Comm_commit((*new_intracomm_ptr));
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /* printf("About to get context id \n"); fflush(stdout); */
-    /* In the multi-threaded case, MPIR_Get_contextid_sparse assumes that the
-     * calling routine already holds the single criticial section */
-    new_context_id = 0;
-    mpi_errno = MPIR_Get_contextid_sparse((*new_intracomm_ptr), &new_context_id, FALSE);
-    MPIR_ERR_CHECK(mpi_errno);
-    MPIR_Assert(new_context_id != 0);
-
-    /* We release this communicator that was involved just to
-     * get valid context id and create true one
-     */
-    mpi_errno = MPIR_Comm_release(*new_intracomm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIR_Comm_create(new_intracomm_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    (*new_intracomm_ptr)->remote_size = (*new_intracomm_ptr)->local_size = new_size;
-    (*new_intracomm_ptr)->rank = -1;
-    (*new_intracomm_ptr)->comm_kind = MPIR_COMM_KIND__INTRACOMM;
-    (*new_intracomm_ptr)->context_id = new_context_id;
-    (*new_intracomm_ptr)->recvcontext_id = new_context_id;
-
-    mpi_errno = create_and_map(comm_ptr, local_high, (*new_intracomm_ptr));
-    MPIR_ERR_CHECK(mpi_errno);
-
-    (*new_intracomm_ptr)->tainted = 1;
-    mpi_errno = MPIR_Comm_commit((*new_intracomm_ptr));
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIR_COMM_KIND__INTERCOMM_MERGE_IMPL);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-
 #endif
